create separate testing package with testing utils
