--- packages/content/__mocks__/fs.cjs
@@ -1,2 +1,0 @@
-const { fs } = require('memfs');
-module.exports = fs;
