--- packages/content/__mocks__/fs/promises.cjs
@@ -1,3 +1,0 @@
-const { fs } = require('memfs');
-module.exports = fs.promises;
-
