+++ packages/content/package.json
@@ -57,6 +57,7 @@
     "strip-json-comments": "^4.0.0"
   },
   "devDependencies": {
+    "@bluecadet/launchpad-testing": "0.1.0",
     "@bluecadet/launchpad-tsconfig": "0.1.0",
     "@types/cli-progress": "^3.11.0",
     "@types/markdown-it": "^12.2.0",
