--- packages/content/test/setup-tests.js
@@ -1,6 +1,0 @@
-import { vi } from 'vitest';
-
-vi.mock('fs');
-vi.mock('fs/promises');
-vi.mock('node:fs');
-vi.mock('node:fs/promises');
