+++ packages/content/vitest.config.js
@@ -4,7 +4,7 @@ export default defineConfig({
 	test: {
 		globals: true,
 		environment: 'node',
-		setupFiles: './test/setup-tests.js',
+		setupFiles: '@bluecadet/launchpad-testing/setup.js',
 		server: {
 			deps: {
 				// inline these deps so that they use the fs mocks
